Dockerfile: run all apt commands at once
Since the image is never squashed, running commands like apt-get clean
in a separate RUN command doesn't actually remove the files from the
final image.

Joining the RUN lines reduces the image size by 23MB.

While at it, set up PATH in the image so that headless-shell is runnable
from within a container. This is useful if the image is used with a
shell or another program, such as when running CI jobs or tests.