change legacy SwiftShader GL to SwANGLE
Legacy SwiftShader GL is deprecated, and the shared libraries on
which headless_shell depends have changed.

The dependencies of headless_shell are found with this command (on the
builder, assuming the directory is "out/Default"):

  strace -o strace.log -f -e trace=open,openat \
    ./headless_shell --no-sandbox --use-gl=angle --use-angle=swiftshader \
    https://get.webgl.org/

Then grep for "out/Default":

  grep -F "out/Default" strace.log

Here is the list of found files:

  - libEGL.so
  - libGLESv2.so
  - libvk_swiftshader.so
  - libvulkan.so.1
  - vk_swiftshader_icd.json

Notes:

1. there is a directory named "swiftshader" containing libEGL.so and
libGLESv2.so, but they are not in use.
2. this is based on 102.0.5005.115, and it seems that the code for GPU
is a fast moving part. We should check frequently.
3. chrome-sandbox is not included in chromedp/headless-shell, so I
removed it. But I'm worried that the builder uses a local modified
build-headless-shell.sh that is not checked into the repository.

References:

https://github.com/chromium/chromium/blob/main/docs/gpu/swiftshader.md
https://bugs.chromium.org/p/chromium/issues/detail?id=1060139
https://bugs.chromium.org/p/chromium/issues/detail?id=765284

Fixes chromedp/chromedp#1073
Fixes chromedp/chromedp#1074